Add basic HAL
This commit creates a bare bone scaffolding for the hardware abstraction
layer. All interaction with architecture specific registers and assembly
has to go there.
In the future, we could add a small emulation layer to enable running
the monitor in x86 or Arm user mode, for things such as running tests or
symbolic execution.