Use a runner script to orchestrate build and runs
This commit adds a runner crate, which contains a script to handle build
steps and QEMU configuration. This lifts a lot of the complexity from
the justfile into that script, and makes the build process much more
flexible.

This new architecture will make it much easier to add new payloads, but
also to set-up integration tests.