Keep track of the number of exits for each virtual context
This commit adds a field to the virtual context to keep track of the
number of exists. This can be seen as a sort of performance monitoring.

On exit we now print the number of exits, so that we can quickly see how
far we go. We are reaching 15 exits on OpenSBI so far, not bad!