Virtualize program counter
Until now the program counter was not properly virtualized, in the sense that
we were relying on the M-mode mepc to keep track of the payload PC value.
This commit adds a virtual PC to the virtual context, which can very simply be
updated from the emulation code. This paves the way for interrupts emulation
and adding support for running an OS alongside the firmware.