Emulate performance-monitoring registers
For now the counters are considered as not implemented, which means that they
exits but are are read-only 0. In the future we could add support for some of
the performance counters by emulating them (i.e. setting the physical counters
to read the counts), but that is not critical for now and  might be
non-trivial.