Enable configuring the maximum number of exits in config.toml
This commit makes it possible to configure the maximum number of payload exits before terminating. This replace the old method with an hardcoded number.

To configure the max number of exits, edit `config.toml` with:

```
[debug]
max_payload_exits = 400
```

Note: for now we can't parse strings into integers at compile time, so we need
to use a runtime function. This will be fixed with
https://github.com/rust-lang/rust/pull/99322 and we should be able to refactor
using a constant at that point.