Add runner option to limit max number of payload exits
The new option overrides the value from the configuration. This is helpful to
ensure there is a maximal exit count configured when running the tests in the
CI, to avoid infinite loops.

If needed in the future we could consider giving priority to the config file,
but at least with the current behavior the tests execute in the same
environment on the CI and locally.