Simple breakpoint test payload
This PR adds a simple payload that causes a breakpoint exception and tries to
handle it. The test fail on the current main, this commit is just there to
demonstrate how to write such a payload and is not intended to be merged.

NOTE: because trap handling is not yet implemented, the payload is not yet
added to `just test`, but can be run with `just run breakpoint`