Use a separate linker script for payloads
Until now payloads used the same linker script as Mirage itself, this wasn't
causing issues because the executable are compiled as PIC, but now that we want
to install trap handlers from the payload it's easier to know the absolute
addresses. This commit creates a new linker script for use by payloads.

One current limitation is that Mirage does not know how to configure the
payload's stack. In the future we should update the `mirage_api` crate to make
it easy for the payloads to bootstrap themselves and find their own stack.