Add missing CSR registers
With the implementation of trapping we want to be able to trap on unknown CSRs
and inject the trap back into the payload. However we can't return an invalid
opcode exception into the payload for CSRs defined as part of the core RISC-V
spec.
This commit moves the panic on known but unimplemented CSRs into the getters
and setters, rather than in the decode function. This will make it possible to
return invalid opcode without violating the spec.

Co-authored-by: Abel Vexina Wilkinson <abel.vexinawilkinson@epfl.ch>