Save trap information on exit
This commit adds a TrapInfo struct which is used to save all trap information
on exit. Those info can then be used to handle payload traps, and wont be
corrupted if a trap occurs in Mirage itself while handling a payload exit.

Using a standardized TrapInfo struct makes it possible to do some cleaning and
to remove some of the instructions to read raw registers, which is definitely a
good thing :)

Co-authored-by: Abel Vexina Wilkinson <abel.vexinawilkinson@epfl.ch>