Add support for medeleg, mideleg, mtinst, and mtval2
The existence of those registers are conditioned to the availability of S-mode
or hypervisor extensions. For now we are always assuming that they don't exist,
but we will need a better way to handle those later on by introducing a concept
of platform supported features (e.g. support for platform misa).

Co-authored-by: Abel Vexina Wilkinson <abel.vexinawilkinson@epfl.ch>