Add support for misa emulation and configuration
This commit adds proper supper for misa emulation, and makes it possible to
toggle S mode support for the virtual misa. It also disable C extensions
(compression) because Mirage assume all payload instructions are 4 bytes long
for now, as well as the H-mode extension because we don't support it right now.

IMPORTANT: this commit does not change the underlying misa when executing the
payload! Thus the payload executes with the host misa, not the virtual misa. We
should implement a misa switch on context switching between the payload and
Mirage for proper emulation.

S mode can be toggled from the configuration file and default to enabled:

```
[platform]
s_mode = true
```