Update Mirage ABI to use ecall
The previous mirage ABI used ecall (with any values in registers) for success,
and WFI for failure. Turns our we will need to emulate ecalls and WFI, so we
need to redefine the ABI.

This commit changes the Mirage ABI to use ecall with a custom EID, and defines
two Mirage functions: success and failure. In the future we can add more
functions under the same Mirage EID, such as printing utilities.