Expose a setup_payload! macro to reduce boilerplate in test payloads
All test payloads need to have their own entry point (in assembly) and their own panic handler. This creates a lot of duplicated code across all of our tests and makes it hard to evolve our test infrastructure. On example is configuring payload stack, which would require to manually update the entry point assembly of all existing payloads.

This commit solves this issue by exposing a `setup_payload!` macro that
configures both the entry point and the panic handler (and a `payload_panic!`
macro for the only payload that needs a custom entry point).

See #18.