Make payloads bootstrap their own stack
Until now Mirage was responsible for configuring the stack of its payloads, so
far using an hard-coded constant in the `Platform` definition.
This commit updates the entry point assembly to use the stack defined by the
linker script, that way Mirage no longer need to configure the Payload stacks.

This change was made super simple thanks to the refactoring that introduced the
`setup_payload!` macro, since all payloads share the same code now.