Make it possible to run unit tests from the main mirage crate
Because the main Mirage crate is a `no_std` and `no_main` binary it does not
support running unit tests out of the box. That is particularly annoying as it
makes it hard to test the internal mirage logic such as the decoder or virtual
context emulation.
We do have integration tests in place, be we definitely need to have unit
testing!

This commit is mostly doing refactoring and Rust/Cargo configuration to make it
possible to compile Mirage in test mode for the host architecture.
Changes include:
- Removing all assembly outside of the `arch` module.
- Adding a new `host` architecture that is empty, but can be compiled on the
  host.
- Add a new "host" feature that can be used for conditional compilation. For
  now the feature needs to be passed in order to build Mirage for the host.
- Split the `mirage_abi` in two, by introducing the `mirage_core` crate. That
  allows for only importing definitions from `mirage_core`, but not the Risc-V
  specific assembly from `mirage_abi`.
- Conditionally makes Mirage `no_std` and `no_main`, so that it can be compiled
  normally as a user-space program for the host.

While doing all of this I tried to reduce tooling (mostly Rust Analyser)
confusion as much as possible. It seems the current version is not that bad,
but feel free to open an issue in case tooling stops to work (like loosing
completion) due to conditional compilation or architecture confusion.

With all those changes, it is now possible to use `cargo check` and `cargo
test`!

Example:

```
cargo check --features host -p mirage
cargo test  --features host -p mirage
```