Add first unit test and run unit tests on CI
Now that Mirage can be compile and run on the host, we can include unit tests
in the main Mirage crate itself. This commit demonstrates how to add simple
unit tests with an example for the decoder.

Starting from this commit, unit tests are checked with `just test`, and thus
will run on the CI.