Disable floating point extensions
This commit disables the F, D, and Q extensions (float, double, and quand).
We are hitting some errors due to floating point registers in OpenSBI, which go
away when disabling floating point support. We will need a special handling for
saving and restoring floating points registers when supporting an OS anyway, so
we decided to disable them for now until proper support is added.