Update CI to remove warning and apt-get failure
The CI got apt-get failures when downloading qemu. I suspect that's because the
index got stale since the image is pretty old (ubuntu 22.04).
To fix this I updated ubuntu-22.04 to ubuntu-latest, that should get us the
latest shiny images, and I added an apt-get update to refresh the index before
downloading qemu.

The CI also emit warnings for the Just action because it's using an older node
version (v16), I upgraded the CI to the newer just action.