Add a log method to the Mirage ABI
This new Mirage ABI call exposes a log interface to the virtualized payload.
For now the main purpose is to expose a simple interface for test payloads to
print, making it easier to debug our own tests.

This commit includes a logger implementation intended for payload use, which
uses the new Mirage call as a backend. This logger implementation always log
all messages (thus incurs a performance cost!) and let Mirage itself apply the
configured filtering.

WARNING: The new logging API accepts a pointer and a size, but does NOT check
if the message lives en memory that can be access by the payload! This cause
security issue, and will need to be resolved in the future when we add a
concept of accessible memory and set-up appropriate memory protections.