On PR run tests on all new commits in the CI
This commit updates the CI to run our tests on all new commits in the CI.
While a big improvement to enforce that all commit passe all tests, this is not
perfect yet as we there might be changes merged between the tests for a PR and
the merge of that PR, in which case the code can break in the meanwhile.
It might  be possible to mitigate that issue using Github's automerge feature
or some other configuration, I suppose?