Add some initial documentation
This documentation is aimed at Mirage developers, rather than users. This is
why the documentation lives in the repo itself rather than on a website.

The documentation is meant to be a living thing! It is strongly encouraged to
document anything that might arise during code review and requires a decision
to be taken, for instance.