Merge csr_id test into csr_ops
With the recent addition of logging support from the payload, we can now merge
some of our integration tests while retaining and simple way to debug failing
tests.
This commit is an example of merging the csr_id test into csr_ops. Both tests
are similar: they tests interactions with the CSR registers. Now the test can
print what failed in the assert, making it easier to debug an eliminating the
need for maintain two independent integration tests.