Add support for running S-mode payload
This commit introduces a new execution mode for direct execution of S-mode and
U-mode payloads. To do so we introduce a concept of execution mode switch, when
the virtual CPU switches from M-mode to S or U-mode or vice versa.
During an execution mode switch we move CSRs from the virtual to physical CPU
appropriately:
- From Firmware to Payload: install virtual S-mode CSRs in physical registers,
  and configure M-mode CSRs to prepare S and U-mode execution.
- From Payload to Firmware: save the content of S-mode CSRs into the virtual CPU
  context, re-initialize their content so that their don't interfere with
  execution of the firmware in U-mode, and prepare U-mode execution.

WARNING: This commit is not exhaustive, some of the CSRs might not yet be
properly updated and propagated from firmware execution context to payload
execution context. For instance, interrupt related CSRs (sip, sie, mip, mie) are
not yet handled.

Co-authored-by: Abel Vexina Wilkinson <abel.vexinawilkinson@epfl.ch>