Add support for downloading artifacts
This commit adds support for downloading and running artifacts.
Artifacts are defined in an artifact manifest (`misc/artifacts.toml`) with a
name and corresponding URL. When calling the runner with a payload that doesn't
exists in the `payloads` directory the runner looks up the artifact manifest and
downloads the artifact if found. Artifacts are downloaded only once until the
artifact manifest gets updates, at which point the runner will download the
artifacts again on demand to ensure binary artifacts versions are changed in
locksteps with the Mirage sources.

Since we now support running OpenSBI, this commit adds the first test relying on
foreign artifacts using the OpenSBI build from
https://github.com/CharlyCst/mirage-artifact-opensbi