Rename "host" feature to "userspace"
This is part of an effort to clarify the terminology. In this context
host can be ambiguous because Mirage itself is a host virtualizing a
guest firmware (using traditional virtualization terminology), but host
here was referring to the host development platform.
The word "userspace" removes all of that ambiguity and clarify the
intent better.