Parse config strings at compile time
This commit makes it possible to parse config strings at compile time.
Unfortunately `usize::from_str_radix` is not yet stable (but should be
soon), so we had to hand-roll a `const` parser. We can fall back to
`from_str_radix` once it gets stabilized as a const function, see:
https://github.com/rust-lang/rust/pull/124941