Add config option to set maximum number of vPMP
This commit adds a config option to set the maximum number of PMP
exposed to the virtual firmware. For clarity, the option is grouped with
the option to expose (or not) S-mode under a new `vcpu` table.
The existing `platform` table has been removed, but will be added back
in the future to configure the target platform (rather than the vCPU).

In case 0 vPMP are exposed there should be no restrictions to payload
memory access, this commit ensures it is the case.