Do not panic on PMP LOCK bit
Previously Mirage was simply panicking when the firmware was trying to
set the LOCK bit in one of the pmpcfg register. It turns out the bit is
WARL, so we can decide to hard-wire it to zero and not panic.

I decided to still emit a warning when the frimware tries to set the
LOCK bit, just so that we don't forget that the feature is not supported
yet.