Better context formatting on trace log level
Previously the context was simply dumped as an unformatted struct when
running with trace log level. This commit structures the log better.
For now not all registers are included (e.g. PMP registers are not for
instance), but we can add them as we go (note that sometime GDB is the
right option and we are better-off not logging too much).