Add support for WFI
This commit adds support for WFI (wait for interrupt) by simply directly
executing the WFI instruction. For no we do not perform additional
checks nor configure any timer, which means that the firmware can put
the core into a perpetual sleep state. This is effectively a potential
denial of service attack, but we do not seek to prevent those for now.