Initial support for hardware detection
This commit adds initial support for hardware capability detection, with
available interrupts as the first use case.
Hardware detection temporarily modifies the state of the CPU, for this
reason it must be done prior to the initialization of CPU state and
prior to running any virtualized firmware.
Hardware capabilities are described in a new HardwareCapability struct.
Because hardware capabilities are local to a core (hart in RISC-V
parlance) they should not be used by other cores. We enforce this by
marking the struct as !Send and !Sync.