Forward hardware capabilities to CSR registers setters
Some of the CSR registers only support a limited number of values which
depends on the hardware characteristics. This was previously handled
through compile-time configuration, but that approach will not work in
heterogeneous environments and will not scale as we increase the set of
supported registers.
Instead this commit refactors the getters and setters trait to expose a
new setter which takes the hardware capabilities as an argument. This
setter is exclusively used for CSRs so far, and allows for checking the
hardware configuration prior to updating the CSR values.