Runner: introduce subcommands
Initially the runner was simply intended to run Mirage, but as the
project moves forward I anticipate the runner will accumulate more and
more functionalities. This already started with the ability to download
and manage artifacts, for instance.
This commit introduces the notion of subcommands for the runner. For now
we support a single subcommand, `run`, which behaves exactly as the
runner did before (without the `run` subcommand).