Initial multi-core support
This commit adds the initial multi-core (or multi-harts) support for
Mirage. This include both changes in the runner tu configure the number
of cores (i.e. harts) and the stack size, and changes to Mirage boot
procedure to properly initialize the per-core stack.
For now we park all the non-boot harts, so that the firmware does not
have more than one hart.