Use a pre-defined configuration for `just test`
This commit extends the runner to take an optional --config argument for
the `run` command. This argument enables specifying the path to a
configuration, and is used to ensure all the tests are done with a fixed
configuration.
Before that change, `just test` would pick up a custom user
configuration (`./config.toml`), which could be different than the
default values used in the CI.