Fine grained logger with variable levels per module
This commits allows to log specific modules at a lower level. For example the global level could be info but you want to have a few modules with trace for debugging purposes. You can add those modules in the toml configuration and logs will be printed at a higher granularity