Add Zephyr artifact
This commit adds a Zephyr artifact, using the artifact built from a new
repository that we operate:
https://github.com/CharlyCst/mirage-artifact-zephyr
The build instructions are inspired from #73 and adapter to run on the
CI.

For now the artifact is not tested as part of the CI, because we don't
have support for running Zephyr yet. It is still useful to have access
to the Zephyr artifacts though, as it makes it simple to track progress
toward running Zephyr.