Automatically detect the available number of PMP registers
Before this patch we had to specify the number of PMPs manually for each
platform, which is fine but makes it harder to port Miralis to other
environments (for instance running Miralis on Miralis).
With this patch Miralis detects automatically the number of available
PMP registers, removing the need for hard-coding such number for each
platform.

Note that because each PMP register needs to be written with a different
assembly instruction the code for poking at the 64 PMP registers is
quite verbose and macro heavy, but I am not sure how to do better
without going too far onto the macro path.


Co-authored-by: Charly Castes <charly.castes@epfl.ch>