counteren, senvcfg and menvcfg store and restore
When swithcing from payload to firware (and opposite) the registers
senvcfg and menvcfg need to be stored and restored by mirage if they are
availables.

counteren register allow reading into various counter. Last three bits
of mcounteren (IR, TM and CY) may need to be set in order to give
the kernel the rights to read associated counters.

Change test csr_ops when write in mcounteren