CI: Check that no big object has been committed
The size of the repository grows linearly over time, a big pitfall is to
accidentally commit large binary files that then needs to be removed.
This PR checks the size of the biggest object committed and rises an
error if any of them is bigger than 200 kb. This should be big enough
for quite some time as long as we only commit text.

Fixes #47