Add support for decoding of some store and load instructions
This commit adds support for decoding some load and store instructions,
namely ld, sd, cld, csd, and csw. Mirage will need to be able to decode
those instruction when adding support for virtual devices, as the load
and stores to virtual devices will need to be emulated.
It is notable that this commit introduces support for emulating some
compressed instructions, therefore we must be careful when incrementing
the PC due to load/store emulation.