Bugfix: set MPP to 0 after MRET
The spec mandates that MPP is reset to the lowest available privilege
mode when executing MRET. For systems without U-mode this corresponds to
M-mode, otherwise U-mode is always the least privileged mode (and we
assume M-mode with Mirage).
The previous implementation was not resetting MPP appropriately, this
patch implements the correct behaviour.

This bug was found using formal methods to compare the Mirage
instruction emulator with the Sail RISC-V model (i.e. the executable
specification).