Rename Mirage to Miralis
Unfortunately Mirage is quite a bit overloaded as a project name. Most
notably there is MirageOS (https://mirage.io/), which itself is already
shadowing the previous MirageOS (https://www.tiwizard.com/downloads/mirage-os-1-2/).
There are also a lot of less related projects and research paper called
Mirage, which overall just creates confusion.
To ensure people do not mistake Mirage with an OCaml unikernel we
decided to change the name of the project. Naming things is hard, there
was a bunch of bikeshed, but we settled on Miralis. Miralis is a short
name, easy an unambiguous to pronounce, and most importantly there is no
related project with similar name.