Flatten firmware directory tree
This commit removes the `src` folder from the various test firmware
directories. It simply updates the `Cargo.toml` file to point into a
`main.rs` Rust file at the root of the directory rather than in `src`.
This change will make it easier to navigate the repo, as it is one less
directory to traverse each time one access a test firmware file.