Add support for choosing platform in config.toml
This patch makes it possible to choose which platform to compile for
from the config.toml file. The platform is simply selected from the
`platform.name` attribute.
This commit adds a `visionfive2.toml` config to prepare for the upcoming
visionfive2 support and check in the CI that said config file is well
formed.