Add support for choosing start_address in config file
This patch makes it possible to choose the start address of Miralis from
the configuration file. This is a requirement to support multiple
platforms, in particular we need it for our ongoing effort to port
Miralis to the VisionFive 2 board.

The solution to change the address is to create a new symbol with the
desired address as value (`_start_address`). This can be done by passing
an argument to the linker through Cargo. Our runner already passes
linker arguments (namely the linker script path), so this already fits
well with our existing tools.