Add initial support for VisionFive2 boards
With this commit Miralis now has (initial) support for the VisionFive2
board! With this commit Miralis can boot on VisionFive2, use the uart
(with a small caveat, see bellow), and jump into the firmare. So far
both OpenSBI (without payload) and Zephyr were tested on the hardware.

For the uart we skip the initialization for now, we assume that the
previous boot stage (U-boot SPL) has already configured the uart
properly. We do busy waiting for a fixed amount of time before writting
the next character, this can be improved in the future.