Implement userspace Arch trait
This commit implements the Architecture trait for userspace, which
allows running more of Miralis code in unit tests (i.e. on the host
architecture as a user-space process).

In the long term we plan to use this new capability to verify more of
Miralis' invariants: we can log which system registers are set (because
all of the Miralis codebase as to go through `Arch` to set system
registers), and verify some properties, such as `medeleg` is always 0,
or that interrupts are properly configured.

To demonstrate the new userspace implementation this commit includes
some unit test to verify some invariants during trap handling.