Bugfix: properly exit with an error when config is invalid
Our runner has a new `check-config` command, but it turns out the
command was only printing the error, not exiting with a non-zero code.
Because of that our CI was passing despite a type in the name of one of
the config file. This patch makes the CI report error properly if one of
the config is broken.