Remove multiplication instruction in stack setup code
This patch removes the `mul` instruction that we used in assembly. We
had to do this because the compiler was emitting errors due to the use
of multiply (but multiply might not be present on the machine).

This commit replaces the multiplication with a loop. Basically, each
hart needs to compute the start of its stack, which is the global stack
start address + stack size * hart_id. This calculation is now done using
a loop, adding one stack size for each hart.