Benchmark configuration and documentation.
Benchmark can now be set using the config.toml file within the table
[benchmark]. Type of benchmarks can also be selected to be run or not.

Benchmark analyzer is now also a library to allow the runner to run
alalysis directly. The executable version allows to run analysis on the
reccord of a run written in a file.

Documentation of just benchmark and just analyze-benchmark commands.
Add benchmark as a test.

Add custom print version for the benchmark.