Refactor benchmark code and add more counters
Code is refactored to make it easier to add a counter.
Two type of counters: first one is basic counters with purpose of
counting occurences of events. The second one aims to measure difference
beetween two events.
The counters shares many properties that are grouped under the Either
enum.

Other properties are recorded like number of world switches or
firmware exits.