Encode read/write width for device accesses
Until this commit we were not making any distinction based on the width
for read and write access. This mean that the width of the access was
basically assumed to be correct by our only driver (the clint), but for
some devices we might not know the expected width in advance (for
instance if we add a pass through device which forward all the accesses
directly).