Improve benchmark firmware organization
Add a way to parametrize benchmark, `benchmark.nb_iter`, to configure
how many time a benchmark should run. This makes it possible to
configure how long a benchmark should run and thus run our benchmarks as
tests with few iterations.
To make it possible to read the configuration from the firmwares the
helpers for reading environment values have been factorized into a crate
(`config_helper`).

Finally the benchmarks have been moved into the `firmware/benchmark`
directory to make it easier to group them together. This required
extending the search path of the runner to `firmware/benchmark` in
addition to `firmware`.