Benchmark analyzer use now csv format
Benchmark can print output either in a human readable way or in csv
format. This can be controlled by a flag in the config and is
overwritten to true when running benchmarks. CSV allows to easily parse
results and merge results of multiple runs.

Modify csr_write_benchmark to be benchmark compatible