Fix SEIE bit overwrite
SEIE bit of mip register was overwritten when context switches by a
wrong value. This value comes from the fact that csrr instruction is
modified for SEIE and output the logical OR of what is in mip and if
there is a signal from an external device (here, it was the PLIC).

We then sync this bit between physical and virtual mip and take care to
not take a wrong value of SEIE: