Add support for loading payloads
This commit adds support for loading custom payloads in QEMU, as well as
support for automatically building them like we do with firmware.
Payloads can be specified in an optional section in the config file
(`[target.payload]`) with a choose start address. When this section is
present the runner will build and load the payload into memory.

Instead of duplicating the firmware infrastructure (such as linker
script, boot assembly, etc...), we unified the payload and firmware to
use the same linker script and bootstrap code. At this occasion we
migrated the firmware to use the same linker script as Miralis,
therefore we only have a single linker script as of this commit, used by
Miralis, the firmware, and the payload.