Remove just benchmark command
The `just benchmark` command runs the benchmarks on QEMU, but we
realized that performance counters can not be used to produce actionable
insights (for instance the instruction retired count does not match what
we measure on real hardware).
Since `just benchmark` can't be used to produce meaningful result we
decided to remove it for the sake of simplification.