Separate build profiles for Miralis and firmware
With this patch the build profile for Miralis and the firmware can be
configured independently. They are configured through two new config
sections: `target.miralis` and `target.firmware`.
The start address has also been moved under those config section, and we
anticipate that in the future the stack size would move there too.