Add support for per-hart stack init for firmware
With this commit firmware stack size is no longer defined in the
corresponding linker script. Now it is defined on a per-hart basis in
the config file, similar to miralis' stack size.
Also the stack_size option in config is moved from "platform" section to
the "target.[miralis/firmware]" section.