Bugfix: forward signals to GDB on Unix platforms
The runner is responsible for spawning the GDB process, because the GDB
command depends on the configuration.
Prior to this patch GDB was spawned as a new process separate from the
runner. This has the unfortunate consequences that GDB does not receive
signals (most notably Ctrl-C) as they are intercepted by the runner
process in the foreground.
To fix this we now start the GDB process with `exec` on Unix platforms,
therefore replacing the runner process by a brand new GDB process in the
foreground. On non-Unix platforms we keep the old (buggy) behavior for
now by simplicity.