Initial support for running Miralis as a Miralis firmware
This commit introduces support for booting Miralis as a Miralis firmware
(in order to prepare for nested virtualization). We actually use
paravitualization and rely on calls to the host for functions such as
logging and exiting, which make it possible to nicely integrate with the
host Miralis and facilitate debugging. This is implemented by
introducing a new "Miralis" platform.

With this commit the nested Miralis does not yet start executing
firmware, it simply stops after initialization.