Avoid MEIE and MSIP to be set in the virtual context
To wake up harts, firmware might use a machine software interrupt (`MSI`)
or am machine external interrupt (`MEI`). These interrupts need to be fetched
from hardware (`mip`) for each virtual read of `vmip`, as they can occur
asynchronously with the execution of the firmware. During a world switch,
we need to take care that these interrupts are not installed in the virtuals
`vmip`, to avoid having an interrupt that can't be cleared by firmware in the
virtual context.