Automatically collect statistics when pushing on main branch
This commit adds a new github workflow that runs when pushing commits on
main. The workflow collects statistics about Miralis, for now binary
size and cold build time, and pushes them to another repository
(https://github.com/CharlyCst/miralis-commit-stats).

The current implementation is a bit fragile, it is kind of hard to write
a somewhat portable and robust bash script. Because the script is not
portable enough it is not tested in the main CI, maybe at some point we
will want to integrate this functionality into the runner.