Add runner command to list artifact
This patch adds a new `artifact` command to the runner that lists the
available artifacts. The command takes an optional `--markdown` argument
that prints the list of artifacts plus their description and metadata in
markdown. The markdown output is intended for auto-generating
documentation, which we will use in the Miralis website.