Set runner as the default package
This commit sets the runner crate as the default package. This makes it
easier to call the runner, as `cargo run --` will invoke the runner
directly. Prior to this commit the command was `cargo run -p runner --`,
which is arguably not as nice.
As a consequence the justfile can be simplified by removing the now
unnecessary `--package runner` flags.

In the future I am thinking about getting rid of the justfile entirely,
and directly invoking the runner for all of our need. This would require
fixing a few issue, such as stable invocations across the git history,
auto-completion, and more test automation, but this would help working
around some of the just limitations (such as the lack for named
arguments).