Introduce proc macro to select a type from an environment variable
This allows to remove the `#[cfg]` guards for the platforms and (in the
future) for the policy modules. The problem with `#[cfg]` guards is that
they don't scale well with the number of choices, and introduce
bug-prone conditional compilation which we only really need for the
architecture-dependant assembly.

The new `select_env` proc macro takes the name of an environment
variable, and then matches its value against a list of possible values.
It then replaces the macro invocation with the patch of the matching
arm.