Hypervisor extension emulation
In the RISC-V architecture, the Hypervisor extension (often referred to as H-extension) introduces a new privilege mode (HS-mode) that provides a virtualized environment. This allows a hypervisor to manage virtual machines, trapping certain privileged instructions and handling them in a way that maintains isolation between the host and guest operating systems. This commit emulates the behavior of this ISA extension in Miralis.