Improve PMP flushing ergonomics
This patch modifies `write_pmp` to return a `#[must_use]` struct that
can be consumed to flush the caches, which is required to apply the
changes in the PMP configuration.
This will make it easier to use the PMP properly, as the compiler will
rise an error if the caches are not flushed after a PMP change.