Detect if H and S extensions at runtime
Currently, the availability of the S and H extensions is determined at compile time. However, we've encountered scenarios where Miralis is compiled with H and S extensions enabled, even on hardware that doesn't support them. This can lead to silent crashes, complicating the debugging process. This commit improves the overall stability of Miralis by addressing this issue.