Bugfix: properly display debug message in runner
The runner was recently updated with a new logging facility, but turns
out it was not displaying debug-level messages even with the `--verbose`
flag.
This commit fixes the issue, and cleans the output a bit more by
removing another non-useful log. The runner now properly displays the
QEMU command when invoked with verbose.