Upgrade toolchain to 1.82
With the new 1.82 release here are a bunch of new features that got
stabilized that are of interest to us. Technically our nightly toolchain
points to rust 1.84, but as we picked a date after the stable 1.82
release we can rely on the features from 1.82 being stable now.

This commit bumps the toolchain and add a few changes to comply with the
new lints, as well as new LLVM target.