Virtualize machine timer interrupt (MTI)
This PR implements proper MTI virtualization. For that purpose it lays
out the foundations of full interrupt virtualization, i.e. it lets
Miralis manage the `mip` bits and inject interrupts rather than relying
on the hardware physical interrupts.

For now all physical M-mode timer interrupts are still assumed to be
caused by the firmware and update the virtual `mip.MTIP` bit. If we need
out-of-band MTI we should implement a mechanism for distinguishing the
cause of a timer interrupt, most likely simply checking if the virtual
`mtimecmp` deadline passed or not.