Ignore mip.SEIE hardware bit on virtual CSR get
The mip.SEIE bit is in fact composed of two different bits, a
hardware-writeable one and another software-writable one. The behavior
of CSR set and clear operations is modified for the particular case of
this bit. To emulate it properly we would need to special case all of
the CSR set and clear functions to ignore the hardware-writeable bit.

Instead of duplicating the whole logic, this commit simply omit the
hardware-writeable bit, making it invisible to the virtualized firmware.
In practice this is not really an issue as we force the external
supervisor interrupts to be delegated, so the firmware should not be in
charge of devices generating such interrupts (because we don't want to
emulate them).

This commit fixes the hangs with the Linux shell payload when the Sstc
exension is not present. This can be tested with the following config
parameters on a sufficiently modern QEMU:

```
[qemu]
machine = "virt"
cpu = "sifive-u54,sstc=false"
```